Unify benchmark files into single benchmark.mjs
- Consolidate extended_benchmark.mjs into benchmark.mjs
- Remove analyze_typograf.mjs (analysis incorporated into main benchmark)
- benchmark.mjs now includes:
  - All 6 competitor packages (typograf, retext-smartypants added)
  - Fair normalization (spaces, dashes, copyright)
  - Performance benchmarks
  - Typograf unique features section
